refactor: unify cart context and harden checkout
