+++ app/api/checkout/route.ts
@@ -1,20 +1,22 @@
-import { NextRequest, NextResponse } from 'next/server';
-// BEFORE (failing)
-
-// AFTER (uses alias)
+import { NextRequest, NextResponse } from 'next/server'
 import { stripe } from "@/lib/stripe"
 
 
 export async function POST(req: NextRequest) {
   try {
-    const body = await req.json().catch(() => ({}));
-    const items = Array.isArray(body?.items) ? body.items : [
-      { name: "Nature's Way Soil — Sample Item", amount: 3999, quantity: 1 }
-    ];
+    if (!stripe) {
+      return NextResponse.json({ error: 'Stripe not configured' }, { status: 500 })
+    }
+
+    const body = await req.json().catch(() => ({}))
+    const items = Array.isArray(body?.items) ? body.items : null
+    if (!items) {
+      return NextResponse.json({ error: 'No items to purchase' }, { status: 400 })
+    }
 
     const origin =
       process.env.NEXT_PUBLIC_SITE_URL ||
-      `${req.nextUrl.protocol}//${req.headers.get('host')}`;
+      `${req.nextUrl.protocol}//${req.headers.get('host')}`
 
     const session = await stripe.checkout.sessions.create({
       mode: 'payment',
@@ -28,10 +30,10 @@ export async function POST(req: NextRequest) {
       })),
       success_url: `${origin}/success?session_id={CHECKOUT_SESSION_ID}`,
       cancel_url: `${origin}/cancel`
-    });
+    })
 
-    return NextResponse.json({ url: session.url }, { status: 200 });
+    return NextResponse.json({ url: session.url }, { status: 200 })
   } catch (err: any) {
-    return NextResponse.json({ error: err.message || 'Server error' }, { status: 500 });
+    return NextResponse.json({ error: err.message || 'Server error' }, { status: 500 })
   }
 }
