+++ pages/_app.tsx
@@ -1,5 +1,5 @@
 import type { AppProps } from 'next/app';
-import { CartProvider } from '../context/CartContext';
+import { CartProvider } from '../lib/cart';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
